requirements.txt
pydantic-ai
yfinance
gradio
python-dotenv
